Добавляет функциональный компонент VideoPlayer и импортирует его в компонент Card
