+++ src/components/card/card.jsx
@@ -1,8 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import VideoPlayer from '../video-player/video-player.jsx';
 
 const Card = ({film, onFilmCardClick, onMouseEnterCard, onMouseLeaveCard}) => {
-  const {id, title, preview} = film;
+  const {id, title, preview, video} = film;
 
   const handleCardEnter = () => onMouseEnterCard(id);
 
@@ -20,6 +21,7 @@ const Card = ({film, onFilmCardClick, onMouseEnterCard, onMouseLeaveCard}) => {
         onClick={() => onFilmCardClick(id)}
       >
         <img src={`img/${preview}`} alt={title} width="280" height="175" />
+        <VideoPlayer video={video} poster={preview} />
       </div>
       <h3
         className="small-movie-card__title"
@@ -38,7 +40,8 @@ Card.propTypes = {
   film: PropTypes.shape({
     id: PropTypes.number.isRequired,
     title: PropTypes.string.isRequired,
-    preview: PropTypes.string.isRequired
+    preview: PropTypes.string.isRequired,
+    video: PropTypes.string.isRequired
   }).isRequired,
   onFilmCardClick: PropTypes.func.isRequired,
   onMouseEnterCard: PropTypes.func.isRequired,
